# alon_sentinel/docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Alon Sentinel API
  version: 1.0.0
  summary: Stable public HTTP contract for Alon Sentinel v1.
  description: |
    Alon Sentinel is an API-first monitoring service. This contract mirrors
    the Axum route surface implemented under `src/api/routes`.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/token:
    post:
      operationId: issueApiToken
      tags: [Auth]
      summary: Issue an API client access token
      security: []
      requestBody: { $ref: "#/components/requestBodies/IssueToken" }
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" } }
  /v1/auth/revoke:
    post:
      operationId: revokeApiToken
      tags: [Auth]
      summary: Revoke the current API access token
      requestBody: { $ref: "#/components/requestBodies/RevokeToken" }
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" }, "403": { $ref: "#/components/responses/Forbidden" } }
  /v1/admin/auth/login:
    post:
      operationId: issueAdminToken
      tags: [Admin Auth]
      summary: Issue an admin access token
      security: []
      requestBody: { $ref: "#/components/requestBodies/IssueAdminToken" }
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" }, "429": { $ref: "#/components/responses/TooManyRequests" } }
  /v1/admin/auth/logout:
    post:
      operationId: logoutAdmin
      tags: [Admin Auth]
      summary: Revoke the current admin access token
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "401": { $ref: "#/components/responses/Unauthorized" } }
  /v1/admin/auth/me:
    get:
      operationId: getAdminMe
      tags: [Admin Auth]
      summary: Get the current admin session
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "401": { $ref: "#/components/responses/Unauthorized" } }
...
  /v1/notifications/channels:
    get:
      operationId: listNotificationChannels
      tags: [Notifications]
      summary: List installation notification channels
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "401": { $ref: "#/components/responses/Unauthorized" }, "403": { $ref: "#/components/responses/Forbidden" } }
    post:
      operationId: createNotificationChannel
      tags: [Notifications]
      summary: Create installation notification channel
      requestBody: { $ref: "#/components/requestBodies/NotificationChannel" }
      responses: { "201": { $ref: "#/components/responses/JsonCreated" }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" }, "403": { $ref: "#/components/responses/Forbidden" } }
  /v1/notifications/channels/{channel_id}:
    parameters: [ { $ref: "#/components/parameters/ChannelId" } ]
    patch:
      operationId: updateNotificationChannel
      tags: [Notifications]
      summary: Update installation notification channel
      requestBody: { $ref: "#/components/requestBodies/NotificationChannel" }
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" }, "403": { $ref: "#/components/responses/Forbidden" }, "404": { $ref: "#/components/responses/NotFound" } }
    delete:
      operationId: deleteNotificationChannel
      tags: [Notifications]
      summary: Delete installation notification channel
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "401": { $ref: "#/components/responses/Unauthorized" }, "403": { $ref: "#/components/responses/Forbidden" }, "404": { $ref: "#/components/responses/NotFound" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
...
  parameters:
...
    ChannelId: { name: channel_id, in: path, required: true, schema: { type: integer, format: int64 } }
...
  requestBodies:
    IssueToken: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/IssueTokenRequest" } } } }
    RevokeToken: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RevokeTokenRequest" } } } }
    IssueAdminToken: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/IssueAdminTokenRequest" } } } }
...
    NotificationChannel: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpsertNotificationChannelRequest" } } } }
...
  responses:
    JsonOk:
      description: Success
      content: { application/json: { schema: { $ref: "#/components/schemas/JsonValue" } } }
    JsonCreated:
      description: Created
      content: { application/json: { schema: { $ref: "#/components/schemas/JsonValue" } } }
...
    BadRequest: { description: Bad request, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Unauthorized: { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Forbidden: { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
    TooManyRequests: { description: Too many requests, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  schemas:
    JsonValue: {}
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { type: string } }
    IssueTokenRequest:
      type: object
      required: [client_id, client_secret]
      properties: { client_id: { type: string }, client_secret: { type: string, writeOnly: true } }
    RevokeTokenRequest:
      type: object
      required: [token]
      properties: { token: { type: string }, revoked_reason: { type: [string, "null"] } }
    IssueAdminTokenRequest:
      type: object
      required: [email, password]
      properties: { email: { type: string, format: email }, password: { type: string, writeOnly: true } }
...
    UpsertNotificationChannelRequest:
      type: object
      required: [channel_type, name, destination, notify_on_failure, notify_on_recovery, is_active]
      properties: { channel_type: { type: string, enum: [webhook, slack, discord, email] }, name: { type: string }, destination: { type: string }, webhook_secret: { type: [string, "null"], writeOnly: true }, notify_on_failure: { type: boolean }, notify_on_recovery: { type: boolean }, is_active: { type: boolean } }
...